chore: rewrite command files to ts
