+++ src/commands/utility/ping.ts
@@ -1,10 +1,10 @@
-import { SlashCommandBuilder } from "discord.js";
+import {ChatInputCommandInteraction, CommandInteraction, SlashCommandBuilder} from "discord.js";
 
 
 export const command = {
   data: new SlashCommandBuilder().setName("ping")
     .setDescription("Replies with Pong!"),
-  async execute(interaction) {
+  async execute(interaction: ChatInputCommandInteraction) {
     await interaction.reply("Pong!");
   },
 }
