--- src/commands/utility/server.js
@@ -1,11 +1,0 @@
-import { SlashCommandBuilder } from 'discord.js';
-
-export const command = {
-    data: new SlashCommandBuilder()
-        .setName('server')
-        .setDescription('Provides information about the server.'),
-    async execute (interaction) {
-        await interaction.reply(`This server is ${interaction.guild.name} and has ${interaction.guild.memberCount} members.`);
-    }
-
-}
