--- src/commands/utility/user.js
@@ -1,11 +1,0 @@
-import { SlashCommandBuilder } from 'discord.js';
-
-export const command = {
-    data: new SlashCommandBuilder()
-        .setName('user')
-        .setDescription('Provides information about the user.'),
-
-    async execute (interaction) {
-        await interaction.reply(`This command was run by ${interaction.user.username}, who joined on ${interaction.member.joinedAt}.`);
-    }
-}
